# Support parallel decompression (pigz)