This PR speed up image (layer) decompression by using parallel gzip (http://github.com/klauspost/pgzip).

In the first revision decompression was done by running `unpigz` external process (similarly to how it's done in moby - https://github.com/moby/moby/pull/35697), but after performing benchmarks it turned out that it's surprisingly slow (I think it's due to external process overhead and moving data back and forth via stdin/stdout).

Here are benchmark results:

1. exec unpigz vs go gzip (unpigz is slower due to overhead)

```
BenchmarkGzipDecompression128-4    	       5	 209942631 ns/op	 975.31 MB/s
BenchmarkGzipDecompression256-4    	       5	 382351460 ns/op	 702.06 MB/s
BenchmarkGzipDecompression512-4    	       2	 550383735 ns/op	 639.45 MB/s
BenchmarkGzipDecompression1024-4   	       1	2960740138 ns/op	 362.66 MB/s

BenchmarkPigzDecompression128-4    	       3	 406130786 ns/op	 330.48 MB/s
BenchmarkPigzDecompression256-4    	       2	 813984409 ns/op	 329.78 MB/s
BenchmarkPigzDecompression512-4    	       1	1625377583 ns/op	 330.31 MB/s
BenchmarkPigzDecompression1024-4   	       1	3258120313 ns/op	 329.56 MB/s
```

2. exec unpigz vs exec gzip (this was expected behavior, pigz is twice faster, but both decompressors are external processes)

```
BenchmarkGzipDecompression64-4     	       3	 463050022 ns/op	 144.93 MB/s
BenchmarkGzipDecompression128-4    	       1	1005415857 ns/op	 133.49 MB/s
BenchmarkGzipDecompression256-4    	       1	2197161346 ns/op	 122.17 MB/s
BenchmarkGzipDecompression512-4    	       1	4510266184 ns/op	 119.03 MB/s
BenchmarkGzipDecompression1024-4   	       1	7745037689 ns/op	 138.64 MB/s

BenchmarkPigzDecompression64-4     	       5	 204790158 ns/op	 327.70 MB/s
BenchmarkPigzDecompression128-4    	       3	 406545824 ns/op	 330.14 MB/s
BenchmarkPigzDecompression256-4    	       2	 811413161 ns/op	 330.82 MB/s
BenchmarkPigzDecompression512-4    	       1	1624756288 ns/op	 330.43 MB/s
BenchmarkPigzDecompression1024-4   	       1	4252662532 ns/op	 252.49 MB/s
```

3. go gzip vs go pgzip

```
BenchmarkGzipDecompression64-4      	      20	  64101550 ns/op	1046.91 MB/s
BenchmarkGzipDecompression128-4     	      10	 147177273 ns/op	 911.95 MB/s
BenchmarkGzipDecompression256-4     	       5	 258219949 ns/op	 839.56 MB/s
BenchmarkGzipDecompression512-4     	       1	1119090817 ns/op	 479.74 MB/s
BenchmarkGzipDecompression1024-4    	       1	2742437092 ns/op	 391.53 MB/s

BenchmarkPgzipDecompression64-4     	      20	  59348248 ns/op	1130.76 MB/s
BenchmarkPgzipDecompression128-4    	      10	 119474318 ns/op	1123.40 MB/s
BenchmarkPgzipDecompression256-4    	       5	 253783084 ns/op	1057.74 MB/s
BenchmarkPgzipDecompression512-4    	       3	 482842690 ns/op	1111.90 MB/s
BenchmarkPgzipDecompression1024-4   	       1	1128599249 ns/op	 951.39 MB/s
```

CC @samuelkarp

Related #1896