@dmcgowan 2018-09-13
Is there any reason not to make this a compilation option and use pgzip by default?