@dmcgowan 2018-09-13
I am seeing much different benchmark results. Are you running these with any other flags to see the benefits of parallelism?

```
$ go test -bench . -benchmem -tags use_go_gzip_decompressor ./archive/compression > /tmp/1
$ go test -bench . -benchmem ./archive/compression > /tmp/2
$ benchcmp /tmp/1 /tmp/2
benchmark                                      old ns/op     new ns/op     delta
BenchmarkDecompression/decompress-64mb-4       10129048      15112975      +49.20%
BenchmarkDecompression/decompress-128mb-4      20902408      28474168      +36.22%
BenchmarkDecompression/decompress-256mb-4      39889272      56016822      +40.43%
BenchmarkDecompression/decompress-512mb-4      84243279      145920711     +73.21%
BenchmarkDecompression/decompress-1024mb-4     163285345     253962614     +55.53%

benchmark                                      old MB/s     new MB/s     speedup
BenchmarkDecompression/decompress-64mb-4       6625.39      4440.48      0.67x
BenchmarkDecompression/decompress-128mb-4      6421.16      4713.67      0.73x
BenchmarkDecompression/decompress-256mb-4      6729.51      4792.05      0.71x
BenchmarkDecompression/decompress-512mb-4      6372.86      3679.20      0.58x
BenchmarkDecompression/decompress-1024mb-4     6575.86      4227.95      0.64x

benchmark                                      old allocs     new allocs     delta
BenchmarkDecompression/decompress-64mb-4       9              35             +288.89%
BenchmarkDecompression/decompress-128mb-4      9              34             +277.78%
BenchmarkDecompression/decompress-256mb-4      9              34             +277.78%
BenchmarkDecompression/decompress-512mb-4      9              34             +277.78%
BenchmarkDecompression/decompress-1024mb-4     9              34             +277.78%

benchmark                                      old bytes     new bytes     delta
BenchmarkDecompression/decompress-64mb-4       41704         4234139       +10052.84%
BenchmarkDecompression/decompress-128mb-4      41705         4233603       +10051.31%
BenchmarkDecompression/decompress-256mb-4      42687         4234991       +9821.03%
BenchmarkDecompression/decompress-512mb-4      43390         4236906       +9664.71%
BenchmarkDecompression/decompress-1024mb-4     45496         4236917       +9212.72%
```