@mxpv 2018-09-13
@dmcgowan that's interesting. It looks like `ioutil.ReadAll` was the bottleneck for gzip reader. With `io.CopyBuffer` it produces much better results.