@klauspost 2018-09-14
@dmcgowan 4440.48MB/s is way over normal decompression speed. Either the content is way too compressible or uncompressible to be a real-world test. 

Typical speeds should be at least an order of magnitude slower.

Also, if you are just copying to `ioutil.Discard` you will not see any significant benefit. In pgzip, the decompression is merely done in an async goroutine ahead of the actual reader consuming it. So if you actually use the data for processing you will see a speedup, otherwise it will just match the built-in decompression.