@mxpv 2018-09-14
@dmcgowan I've reworked benchmark to make it a bit closer to real world scenario.
Here are my results:
```
$ go test -bench . -tags use_go_gzip_decompressor > /tmp/1
$ go test -bench .  > /tmp/2
$ benchcmp /tmp/1 /tmp/2
benchmark                                     old ns/op      new ns/op      delta
BenchmarkDecompression/decompress-16mb-4      102192219      83716965       -18.08%
BenchmarkDecompression/decompress-32mb-4      203506915      164458274      -19.19%
BenchmarkDecompression/decompress-64mb-4      406238786      333350524      -17.94%
BenchmarkDecompression/decompress-128mb-4     807736530      652771252      -19.19%
BenchmarkDecompression/decompress-256mb-4     1618033941     1324395349     -18.15%

benchmark                                     old MB/s     new MB/s     speedup
BenchmarkDecompression/decompress-16mb-4      186.82       228.05       1.22x
BenchmarkDecompression/decompress-32mb-4      187.63       232.18       1.24x
BenchmarkDecompression/decompress-64mb-4      187.99       229.09       1.22x
BenchmarkDecompression/decompress-128mb-4     189.09       233.98       1.24x
BenchmarkDecompression/decompress-256mb-4     188.79       230.65       1.22x
```