@mxpv 2018-09-14
> I am curious if that was similarly the bottleneck for the external process approach. 

I ran `pigz` with new benchmark and got awesome results. So I'm going to update this PR to use `pigz`

```
$ go test -bench .
pkg: github.com/containerd/containerd/archive/compression
BenchmarkGzipDecompression/gzip-32mb-4         	       5	 205370698 ns/op	 185.92 MB/s
BenchmarkGzipDecompression/gzip-64mb-4         	       3	 404389060 ns/op	 188.84 MB/s
BenchmarkGzipDecompression/gzip-128mb-4        	       2	 802574271 ns/op	 190.30 MB/s
BenchmarkGzipDecompression/gzip-256mb-4        	       1	1637217618 ns/op	 186.58 MB/s

BenchmarkPigzDecompression/pigz-32mb-4         	      20	 101242086 ns/op	 377.15 MB/s
BenchmarkPigzDecompression/pigz-64mb-4         	      10	 202710945 ns/op	 376.73 MB/s
BenchmarkPigzDecompression/pigz-128mb-4        	       3	 386691558 ns/op	 394.98 MB/s
BenchmarkPigzDecompression/pigz-256mb-4        	       2	 771113624 ns/op	 396.14 MB/s
```