@mxpv 2018-09-17
I use this in detectPigz



On Sep 17, 2018, at 8:51 AM, Michael Crosby <notifications@github.com<mailto:notifications@github.com>> wrote:


@crosbymichael commented on this pull request.

________________________________

In archive/compression/compression.go<https://github.com/containerd/containerd/pull/2640#discussion_r218123378>:

> @@ -37,6 +43,16 @@ const (
        Gzip
 )

+const disablePigzEnv = "CONTAINERD_DISABLE_PIGZ"


I don't think this is used anywhere

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub<https://github.com/containerd/containerd/pull/2640#pullrequestreview-155992690>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AA00Nowzm6fBpXcgZbZjKU746R1iaPqeks5ub8T7gaJpZM4WmWl8>.