@crosbymichael 2018-09-17
Can you squash your commits? I'm not a fan of the merge commit, it's better to do a rebase on master, not a merged.  After a squash this LGTM