@dmcgowan 2018-09-17
I am going to add this to 1.2 milestone, we should have this in the release. There are still a few changes I want to see but could be done after we merge this
- Integrate with our daemon configuration, a plugin might not make sense since this could be client side
- Avoid `init` functions which do file reads, this could be replaced with a `sync.Once`
- Put the entire benchmark in a gist or break the external dependency